+++ sageopt/__init__.py
@@ -18,10 +18,13 @@ from sageopt import relaxations
 from sageopt import symbolic
 from sageopt.symbolic import signomials
 from sageopt.symbolic import polynomials
+from sageopt.symbolic import elfs
 
 from sageopt.relaxations import local_refine, local_refine_polys_from_sigs
 from sageopt.relaxations import infer_domain
 from sageopt.relaxations import sage_feasibility, sage_multiplier_search
+
+from sageopt.symbolic.elfs import Elf, spelf
 
 from sageopt.symbolic.signomials import Signomial, standard_sig_monomials, SigDomain
 from sageopt.relaxations import sig_relaxation, sig_constrained_relaxation